refactor: extract handlers for core resources
